+++ src/components/news/hooks/useNewsFilter.ts
@@ -3,18 +3,16 @@ import { useAppDispatch, useAppSelector } from "../../../redux/hooks";
 // import { MouseEvent } from "react";
 import { changeNewsOption } from "../../../redux/news/newsSortSlice";
 
+
 export const useNewsFilter = () => {
   const dispatch = useAppDispatch();
 
-  const { newsSortState } = useAppSelector(
-    (state: RootState) => state.newsSorts
-  );
   const reportOptionToAPI = (NewsSortoption: string): void => {
     dispatch(changeNewsOption(NewsSortoption));
   };
 
   return {
-    newsSortState,
+    // newsSortState,
     reportOptionToAPI,
   };
 };
