+++ src/components/news/hooks/useNewsFormat.ts
@@ -2,22 +2,27 @@ import type { RootState } from "../../../redux/store";
 import { useAppDispatch, useAppSelector } from "../../../redux/hooks";
 import { MouseEvent } from "react";
 import { album, text } from "../../../redux/news/newsformatSlice";
+import { useState } from "react";
 
 export const useNewsFormats = () => {
   const dispatch = useAppDispatch();
 
   const { NewsFormats } = useAppSelector((state: RootState) => state.formats);
-
+  const [focused, setFocused] = useState<boolean>(true);
   const changeToImgFormat = (event: MouseEvent<HTMLElement>): void => {
     dispatch(album());
+    setFocused(!focused);
   };
 
   const changeToTextFormat = (event: MouseEvent<HTMLElement>): void => {
     dispatch(text());
+    setFocused(!focused);
   };
   return {
     NewsFormats,
     changeToTextFormat,
-    changeToImgFormat
+    changeToImgFormat,
+    setFocused,
+    focused
   };
 };
