+++ src/components/news/main/NewsList.tsx
@@ -1,25 +1,18 @@
 import React from "react";
 import FormatView from "./cardList/FormatView";
 import styled from "@emotion/styled";
-import { useNewsFilter } from "./../hooks/useNewsFilter";
-
+import { useNewsSorts } from "../hooks/useNewsSorts";
 const NewsList = () => {
-  const { newsSortState } = useNewsFilter();
-
-  // useEffect( ()=>{
-  //   if(newsSorts === 인기순 쿼리) {
-  //     dispatch()
-  //   }
-  // },[])
-
+  const { newsSortState, newsCurOption } = useNewsSorts();
+  console.log(newsSortState);
   return (
     <ViewFilter>
       {(function viewList() {
-        if (newsSortState === "top") {
+        if (newsSortState === "top" && newsCurOption === "정렬순") {
           return <FormatView />;
-        } else if (newsSortState === "latest") {
+        } else if (newsSortState === "latest" && newsCurOption === "최신순") {
           return <FormatView />;
-        } else if (newsSortState === "populer") {
+        } else if (newsSortState === "populer" && newsCurOption === "인기순") {
           return <FormatView />;
         } else {
           return null;
