+++ src/components/news/main/options/index.tsx
@@ -1,17 +1,20 @@
 import styled from "@emotion/styled";
 import React from "react";
-import SortOptionContainer from "./SortOptionContainer";
 import CommonContainer from "./../../../layout/CommonContainer";
-import { useNewsFormats } from './../../hooks/useNewsFormat';
+import { useNewsFormats } from "./../../hooks/useNewsFormat";
+import SortOptionList from "./SortOptionList";
+import { useState } from "react";
 const Options = () => {
-  const { changeToTextFormat, changeToImgFormat } = useNewsFormats();
+  const { changeToTextFormat, changeToImgFormat, focused } = useNewsFormats();
+
+  console.log(focused);
   return (
     <CommonContainer>
       <Features>
-        <SortOptionContainer />
+        <SortOptionList />
         <TextSizeControl onClick={() => {}} />
-        <ImageFormatViewBtn onClick={changeToImgFormat} />
-        <TextFormatViewBtn onClick={changeToTextFormat} />
+        <ImageFormatViewBtn onClick={changeToImgFormat} focused={focused} />
+        <TextFormatViewBtn focused={focused} onClick={changeToTextFormat} />
       </Features>
     </CommonContainer>
   );
@@ -35,20 +38,30 @@ const TextSizeControl = styled.div`
   background-repeat: no-repeat;
   cursor: pointer;
 `;
-
-const ImageFormatViewBtn = styled.div`
+type Props = {
+  focused: boolean;
+};
+const ImageFormatViewBtn = styled.div<Props>`
   width: 40px;
   height: 40px;
-  background-image: url("images/icon-Grid-filled.svg");
   background-size: cover;
   background-repeat: no-repeat;
   cursor: pointer;
+  background: ${({ focused }) =>
+      focused
+        ? "url(images/icon-Grid-filled.svg)"
+        : "url(images/icon-Grid.svg)"}
+    no-repeat 4.5%;
 `;
 
-const TextFormatViewBtn = styled.div`
+const TextFormatViewBtn = styled.div<Props>`
   width: 40px;
   height: 40px;
-  background-image: url("images/icon-Grid-list.svg");
+  background: ${({ focused }) =>
+      focused
+        ? "url(images/icon-Grid-list.svg)"
+        : "url(images/icon-hamburger-button.svg)"}
+    no-repeat 4.5%;
   background-size: cover;
   background-repeat: no-repeat;
   cursor: pointer;
