--- src/components/news/options/SortOptionContainer.tsx
@@ -1,11 +1,0 @@
-import { useState } from "react";
-import SortOptionList from "./SortOptionList";
-
-const SortOptionContainer = () => {
-  const [newsCurOption, setNewsCurOption] = useState<String>("정렬순");
-  return (
-    <SortOptionList newsCurOption={newsCurOption} setNewsCurOption={setNewsCurOption} />
-  );
-};
-
-export default SortOptionContainer
