+++ src/components/news/tabs/NewsTabList.tsx
@@ -1,16 +1,20 @@
 import styled from "@emotion/styled";
-import React, { useState, DragEvent  } from "react";
-import { useNewsList } from "./../hooks/useNewsList";
-const NewsTabList = () => {
-  const { userSavedKeywords } = useNewsList();
-  const [keywordList, setKeywordlist] = React.useState(userSavedKeywords);
-  const [start, setStart] = useState<any>(null); //시작 위치
-  const [currentTab, setCurrentTab] = useState(0);
+import { useNewsTabList } from "../hooks/useNewsTabList";
 
+const NewsTabList = () => {
+  const {
+    keywordList,
+    currentTab,
+    selectMenuHandler,
+    dragstart,
+    dragOver,
+    dragEnd,
+    drop
+  } = useNewsTabList();
   return (
     <Wrap>
       <TabList>
-        {keywordList.map((list:any, index:number) => (
+        {keywordList.map((list: any, index: number) => (
           <li
             key={index}
             data-position={index}
@@ -33,46 +37,6 @@ const NewsTabList = () => {
       </TabList>
     </Wrap>
   );
-
-  function selectMenuHandler(index: number) {
-    setCurrentTab(index);
-  }
-
-  function dragOver(e: any) {
-    e.preventDefault();
-  }
-
-  function dragstart(e: any) {
-    const target = e.target as HTMLElement;
-    setStart(target);
-    console.log(target);
-    target.classList.add("grabbing");
-    e.dataTransfer.effectAllowed = "move";
-    e.dataTransfer.setData("text/html", `${target}`);
-  }
-
-  function dragEnd(e: any) {
-    const target = e.target as HTMLElement;
-    target.classList.remove("grabbing");
-    e.dataTransfer.dropEffect = "move"; //drag effeect
-  }
-
-  function drop(e: any) {
-    const target = e.target as HTMLElement;
-    target.classList.remove("grabbing");
-    target.classList.toggle("active");
-    let startPosition: number = Number(start.dataset.position);
-    let endPosition: number = Number(target.dataset.position);
-    console.log(endPosition);
-    let newKeywordList = [...keywordList];
-    newKeywordList[startPosition] = newKeywordList.splice(
-      endPosition,
-      1,
-      newKeywordList[startPosition]
-    )[0];
-
-    setKeywordlist(newKeywordList);
-  }
 };
 
 export default NewsTabList;
